Task no. 1 Simple Chatbot
This is Task No. 1 for my AI Internship from CodeSoft. I learned the working of natural language processsing in this task. It was fun doing this task as i came to know about many things now NLP and chatbots respond to out messeges. I will be trying to build more advanced version of this chatbot in future and I am exited for this project.